add crud recipes add cd vaforite

+++ controllers/recipe.js
@@ -2,10 +2,56 @@ const db = require("../models");
 const crypto = require("../utils/crypto");
 const errorHandler = require("../utils/errorHandler");
 const validation = require("../utils/validation");
+const Sequelize = require("sequelize");
+const fs = require("fs");
+const config = require("../config");
+
+const fsPromised = fs.promises;
+
+const Op = Sequelize.Op;
 
 const allRecipe = async (req, res) => {
   try {
-    const recipes = await db.Recipe.findAll();
+    const {
+      timeFrom,
+      timeTo,
+      difficulty,
+      sortField,
+      sortOrder,
+      search,
+    } = req.query;
+    let options = { where: {} };
+    console.log("sortOrder", sortOrder);
+
+    if (search) {
+      // options.where.title = { [Op.like]: "%" + search.toLowerCase() + "%" };
+      const title = Sequelize.where(
+        Sequelize.fn("LOWER", Sequelize.col("title")),
+        "LIKE",
+        "%" + search.toLowerCase() + "%"
+      );
+      const description = Sequelize.where(
+        Sequelize.fn("LOWER", Sequelize.col("description")),
+        "LIKE",
+        "%" + search.toLowerCase() + "%"
+      );
+
+      options.where = {
+        [Op.or]: [title, description],
+      };
+    }
+    if (difficulty) {
+      options.where.difficulty = difficulty;
+    }
+    if (timeFrom && timeTo) {
+      options.where.cookingTime = { [Op.between]: [timeFrom, timeTo] };
+    }
+    if (sortField) {
+      options.order = [[sortField, sortOrder]];
+    }
+    console.log("options", options);
+
+    recipes = await db.Recipe.findAll(options);
 
     res.json(recipes);
   } catch (e) {
@@ -18,7 +64,7 @@ const findOne = async (req, res) => {
   try {
     const { id } = req.params;
 
-    let recipe = await db.Recipe.findAll({
+    let recipe = await db.Recipe.findOne({
       where: { id },
       include: { model: db.User, as: "user", attributes: ["id", "email"] },
     });
@@ -36,13 +82,21 @@ const findOne = async (req, res) => {
 const createRecipe = async (req, res) => {
   try {
     const { authorId, title, description, difficulty, cookingTime } = req.body;
+    const { file } = req;
+
     let userId = req.user.id;
+    let image;
     if (authorId && req.user.role === "admin") {
       userId = authorId;
     }
 
+    if (file) {
+      image = file.path.replace("public/", "");
+    }
+
     let recipe = new db.Recipe({
       authorId: userId,
+      image: image,
       title,
       description,
       difficulty,
@@ -64,6 +118,12 @@ const deleteRecipe = async (req, res) => {
   try {
     const id = req.params.id;
 
+    let recipe = await db.Recipe.findOne({ where: { id } });
+
+    if (!(req.user.role === "admin" || req.user.id === recipe.authorId)) {
+      return res.status(403).send({ message: "Forbidden" });
+    }
+
     await db.Recipe.findByIdAndDelete(id);
 
     res.sendStatus(204);
@@ -80,7 +140,11 @@ const updateRecipe = async (req, res) => {
     const { id } = req.params;
     const { title, description, difficulty, cookingTime } = req.body;
 
-    let recipe = await db.Recipe.findAll({ where: { id } });
+    let recipe = await db.Recipe.findOne({ where: { id } });
+
+    if (!(req.user.role === "admin" || req.user.id === recipe.authorId)) {
+      return res.status(403).send({ message: "Forbidden" });
+    }
 
     if (title) {
       recipe.title = title;
@@ -108,10 +172,45 @@ const updateRecipe = async (req, res) => {
   }
 };
 
+const uploadImg = async (req, res) => {
+  let { file } = req;
+  let recipeId = req.body.id;
+
+  if (!recipeId) {
+    res.status(404).send();
+  }
+
+  if (!file) {
+    res.status(400).send({ field: "inputFile", message: "incorrect file" }); //
+  }
+
+  let recipe = await db.Recipe.findOne({ where: { id: recipeId } });
+
+  if (!(req.user.role === "admin" || req.user.id === recipe.authorId)) {
+    return res.status(403).send({ message: "Forbidden" });
+  }
+
+  if (recipe.image) {
+    await fsPromised
+      .unlink(
+        recipe.image.replace(`${config.addressServer}`, "public/") //file removed
+      )
+      .catch(() => null);
+  }
+
+  recipe.image = file.path.replace("public/", "");
+
+  await db.Recipe.update({ image: recipe.image }, { where: { id: recipe.id } });
+
+  recipe = recipe.toJSON();
+  res.send(recipe);
+};
+
 module.exports = {
   allRecipe,
   createRecipe,
   findOne,
   deleteRecipe,
   updateRecipe,
+  uploadImg,
 };
